requirements.txt
py-clob-client
websocket-client
